Big uppdate

+++ Etrx.Domain/Models/Contest.cs
@@ -10,7 +10,7 @@ public class Contest
         bool frozen, 
         int durationSeconds,
         DateTime? startTime, 
-        int? relativeTimeSeconds, 
+        long? relativeTimeSeconds, 
         string? preparedBy, 
         string? websiteUrl,
         string? description, 
@@ -43,40 +43,22 @@ public class Contest
     }
 
     public int ContestId { get; set; }
-
     public string Name { get; set; } = string.Empty;
-
     public string Type { get; set; } = string.Empty;
-
     public string Phase { get; set; } = string.Empty;
-
     public bool Frozen { get; set; }
-
     public int DurationSeconds { get; set; }
-
     public DateTime? StartTime { get; set; }
-
-    public int? RelativeTimeSeconds { get; set; }
-
+    public long? RelativeTimeSeconds { get; set; }
     public string? PreparedBy { get; set; }
-
     public string? WebsiteUrl { get; set; }
-
     public string? Description { get; set; }
-
     public int? Difficulty { get; set; }
-
     public string? Kind { get; set; }
-
     public string? IcpcRegion { get; set; }
-
     public string? Country { get; set; }
-
     public string? City { get; set; }
-
     public string? Season { get; set; }
-
     public bool Gym { get; set; }
-
     public IQueryable<Problem> Problems { get; set; } = null!;
 }
